finish save

+++ video-api/src/main/java/pet/by/ishangulyev/videoapi/entity/Video.java
@@ -18,6 +18,6 @@ public class Video {
     @DocumentReference(collection = "videoFiles")
     private VideoFile videoFile;
     private long length;
-    private long size;
+    private double size;
     private LocalDateTime uploadDate;
 }
